Moved the CoeffsTable in its own module
